third_party/blink/web_tests/dom/node/DOMNodeRemovedEvent-expected.txt

This test checks that DOMNodeRemovedEvent is emitted once (and only once).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS count is 1
PASS successfullyParsed is true

TEST COMPLETE
